app/requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib
pytest
python-dotenv
tortoise-orm
pyjwt
pytz
python-multipart
asyncpg